Move tests to common folder and add type to register method
